pdf_query_bot/requirements.txt
cassio 
datasets 
langchain==0.2.2
tiktoken
google
langchain-google-genai
langchain_community
python-dotenv
streamlit
PyPDF2